Factor out modeling of a "buffer" page to a separate header file.
