+++ src/stream/ByteQueue.h
@@ -41,6 +41,7 @@
 #ifndef DECOMPRESSOR_SRC_STREAM_QUEUE_H
 #define DECOMPRESSOR_SRC_STREAM_QUEUE_H
 
+#include "stream/Page.h"
 #include "stream/RawStream.h"
 
 #include <cstring>
@@ -51,75 +52,6 @@
 namespace wasm {
 
 namespace decode {
-
-class Page {
-  Page(const Page &) = delete;
-  Page &operator=(const Page &) = delete;
-public:
-  static constexpr size_t SizeLog2 = 12;
-  static constexpr size_t Size = 1 << SizeLog2;
-  static constexpr size_t Mask = Size - 1;
-
-  // Page index associated with address in queue.
-  static constexpr size_t index(size_t Address) {
-    return Address >> Page::SizeLog2;
-  }
-
-  // Returns address within a Page that refers to address.
-  static constexpr size_t address(size_t Address) {
-    return Address & Page::Mask;
-  }
-
-  Page(size_t MinAddress)
-      : Index(Page::index(MinAddress)), MinAddress(MinAddress),
-        MaxAddress(MinAddress) {
-    std::memset(&Buffer, Page::Size, 0);
-  }
-
-  void lock() { ++LockCount; }
-
-  void unlock() {
-    assert(LockCount >= 1);
-    --LockCount;
-  }
-
-  bool isLocked() const { return LockCount > 0; }
-
-  size_t spaceRemaining() const {
-    return
-        (MinAddress + Page::Size == MaxAddress)
-        ? 0
-        : (Page::Size - (MaxAddress & Page::Mask));
-  }
-
-  size_t getMinAddress() const {
-    return MinAddress;
-  }
-
-  size_t getMaxAddress() const {
-    return MaxAddress;
-  }
-
-  void setMaxAddress(size_t NewValue) {
-    MaxAddress = NewValue;
-  }
-
-  void incrementMaxAddress(size_t Increment=1) {
-    MaxAddress += Increment;
-  }
-
-  // The contents of the page.
-  uint8_t Buffer[Page::Size];
-  // The page index of the page.
-  size_t Index;
-  Page *Last = nullptr;
-  Page *Next = nullptr;
-protected:
-  // Note: Buffer address range is [MinAddress, MaxAddress).
-  size_t MinAddress;
-  size_t MaxAddress;
-  size_t LockCount = 0;
-};
 
 class ByteQueue {
   ByteQueue(const ByteQueue &) = delete;
